+++ changes.md
@@ -3,6 +3,7 @@
 ## Version 4.1.0: Quality of life
 
 - It is no longer possible to set a stop time if the previously set stop time has passed. Also added unit tests for this feature.
+- The application will now give you an approximate size of the file it is writing to, rounded up to the nearest 10 MB, in the status messages that it produces.
 
 ## Version 4.0.0: Many, many changes