+++ changes.md
@@ -2,6 +2,7 @@
 
 ## Version 4.1.0: Quality of life
 
+- Each ev42 writer module instance will now publish the number of events written to file.
 - It is no longer possible to set a stop time if the previously set stop time has passed. Also added unit tests for this feature.
 
 ## Version 4.0.0: Many, many changes